Finished fixing bugs now trying to fix error where the MAC address is set to 0

+++ src/ip/ipv4.h
@@ -9,8 +9,8 @@ typedef struct
 {
     uint8_t version_ihl; // Version (4 bits) + header length (4 bits)
     uint8_t tos;
-    uint8_t total_length;
-    uint8_t identification;
+    uint16_t total_length;
+    uint16_t identification;
     uint16_t flags_fragment_offset;
     uint8_t ttl;
     uint8_t protocol;
